requirements.txt
annotated-doc==0.0.4
anyio==4.12.1
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
faiss-cpu==1.13.2
filelock==3.24.2
Flask==3.1.2
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mlx==0.30.6
mlx-lm==0.30.7
mlx-metal==0.30.6
mpmath==1.3.0
mss==10.1.0
networkx==3.6.1
numpy==2.4.2
packaging==26.0
pillow==12.1.1
protobuf==6.33.5
Pygments==2.19.2
pypdf==6.7.0
pytesseract==0.3.13
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.2
sentencepiece==0.2.1
setuptools==82.0.0
shellingham==1.5.4
soupsieve==2.8.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.1.0
typer==0.23.1
typer-slim==0.23.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
